initial commit
setting up application
* set up a couple of versions constants
to append that to any external txt files or JavaScript files. And when being incremented, that will force most browsers to bring the new version down. And it just saves the trouble of having to clear the cache when things aren't working the way that expect them to
* create a type config.
this will hold configuration information for our application.<br>
  1. port in what port do I want to listen?
  2. what environment are we in ? production or development
  3. API: What URL do I call to my backend API.
  4. DB a struct has one field How do we connect to the database?
  5. stripe: has the stripe info
  * Create a type call application
     config of type config that we just defined.
     couple of loggers
     template cache
     version