Feat  Start Add Authentication use Stateful tokens
create a login route, handler, and template
How Authentication works
. front-end use session Auth
for a website that has log in and doesn't have an API, for example, we actually can use sessions, will allow the user to fill out a login form, will have some handler that validates the username and password and decides whether that person can log in if that person can log in set some kind of sessional variable like authenticated is equal to true or whatever the case may be.And then until that session expires or the user logs out, they can just do whatever they want.

. back-end use tokens
stateful tokens:  these are tokens that we actually validate the user. store the validation, the token or a hash of the token and an expiry date in the database.
And we can manage the tokens that are valid any time f we want to, for example, invalidate a given user, we simply remove that entry from the database, or we market has expired. And the next time they make a request, they can't do anything.