Feat strict pages only for auth users
Verification for virtual terminal so, that when some request to that page, should be verified to be allowed.
send the token to the back-end, verify if so, display the page
this page for only authenticated users can use.
How?
allow authentication to the back end, also authenticate through the front end and store sessional variable, which is being checked on every request to see if user authenticated or not
-- checkAuth
. check to see if local storage has the item
. get the item token
. if it does not exist then redirect to log in
. if it does exist :
.  we want to actually get the token from storage
. create a headers variable const my headers for fetch

-- create is-authenticated route and CheckAuthentication handler
. validate the token, and get associated user
take the information ` myHeaders.append("Authorization", "Bearer " + token)`  received from the end user. And that comes from based layout.  and we're sending that token back to the server to check if token is valid

-- creat stub function authenticateToken
hand it the request, the request has header Authorization to check if there is an error, if is valid then go to the database and get the user based on userID from tokens table by token_hash
. extract the token from the authorization header
grab the necessary header authorization, make sure that it exists.
make sure that the content of the authorization header consists of the word Bearer, followed by a space followed by some string that is exactly 26 characters long.
if we get past all of that, then we can try to get the user from the tokens table from the database