Feat implement signed links for the email message
want to be able to send a link to the front end that will show a form that allows people to choose
a new password. And of course, we want that to be absolutely secure.  we don't want to make that open to just anyone.
 need some means of securing that to make sure that the only people who have access to the link are people who should have access to the link.
there's a number of ways of doing this.
- One really common method is to append a unique token to the email or to give them a token. They have to enter into a form or something like that. you look up that token in the database and verify that it's valid
- use signed email, emails which have an actual hash or code in the url itself that will ensure that url simply can not be changed, also make sure that an email expires after a set length of time.

* get external pkg go-alone
* create  urlSigner internal pkg to use the functionality from go-alone pkg
* need secret key, and use command line flag to define secret key
* need in the back-end some reference of knowing what the url to the front end is
use the urlSigner pkg to in the api handler to sign a valid email

user can enter whatever email address he wants on the front end.
need to make sure we have that user before to send the json back