Feat pagination for the orders
chunk the data up, to break it into manageable slices of data
create a GetAllOrdersPaginated db func
create a GetAllSubscriptionsPaginated db func
need two things:
how many records for a given query? orders per page
what page are we currently on?
offset: tell how far from the beginning of the results we should offset what we are looking for, and the pageSize tells the limit for the query
update the handlers
update the all-sales page template